backend/requirements.txt
# ---- Core Backend ----
fastapi
uvicorn[standard]
pydantic
python-dotenv


pymilvus
sentence-transformers
groq


# ---- Database ----
sqlalchemy
psycopg2-binary
alembic

# ---- Scraping ----
requests
beautifulsoup4
lxml 
pandas 

# ---- LLM + LangChain ----
langchain
langchain-community
langchain-core
langchain-groq
langchain-google-genai
pymilvus

# ---- Utils ----
httpx
tenacity
 

